# openapi.yaml
openapi: 3.0.3
info:
  title: Kioskin API
  description: |
    Локальный HTTP API для управления оверлей-окном Kioskin на Windows.

    Приложение отображает текст в полупрозрачном popup-окне на выбранном мониторе.
    API позволяет обновлять содержимое, сбрасывать состояние, получать список мониторов
    и корректно завершать работу.
  version: 1.0.0
  contact:
    name: Kioskin

servers:
  - url: http://localhost:8282
    description: Локальный сервер (порт зашит в приложении)

tags:
  - name: window
    description: Управление содержимым и прозрачностью окна
  - name: system
    description: Системная информация и жизненный цикл приложения

paths:
  /health:
    get:
      tags: [system]
      summary: Проверка доступности
      description: Возвращает статус работоспособности HTTP-сервера.
      operationId: getHealth
      responses:
        "200":
          description: Сервис доступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                version: v1.0.0
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /monitors:
    get:
      tags: [system]
      summary: Список мониторов
      description: |
        Возвращает прямоугольники всех подключённых мониторов.
        Индексы соответствуют полю `monitor_index` в `config.json`.
      operationId: listMonitors
      responses:
        "200":
          description: Список мониторов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitorInfo"
              example:
                - index: 0
                  left: 0
                  top: 0
                  right: 1920
                  bottom: 1080
                  width: 1920
                  height: 1080
                - index: 1
                  left: 1920
                  top: 0
                  right: 3840
                  bottom: 1080
                  width: 1920
                  height: 1080
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /update:
    post:
      tags: [window]
      summary: Показать список покупок
      description: |
        Принимает **весь список товаров** одним запросом и отображает его в
        окне в виде выровненной таблицы (Наименование, Кол-во, Цена, Сумма) с
        итоговой строкой ИТОГО.

        - Тело — JSON `{"goods": [ ... ]}`; каждый элемент строго с полями
          `name`, `quantity`, `price`, `sum`.
        - Список **заменяется** целиком присланным (не накапливается).
        - Пустой `goods` (`[]`) допустим — показывает пустую таблицу.
        - Окно становится полностью непрозрачным (alpha = 255).

        Очистить окно — `GET /clear`.
      operationId: updateWindow
      requestBody:
        required: true
        description: Список товаров чека
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseList"
            example:
              goods:
                - name: "Молоко 3.2% 950мл"
                  quantity: 2
                  price: 89.90
                  sum: 179.80
                - name: "Вода минеральная с газом 1.5 л"
                  quantity: 1
                  price: 52.90
                  sum: 52.90
      responses:
        "200":
          description: Список показан
          content:
            text/plain:
              schema:
                type: string
              example: "Список обновлён (позиций: 2)"
        "400":
          description: Некорректный JSON, нет поля goods или в позиции нет обязательных полей
          content:
            text/plain:
              schema:
                type: string
              examples:
                noGoods:
                  summary: Нет поля goods
                  value: "отсутствует обязательное поле goods (список товаров)"
                missing:
                  summary: В позиции нет полей
                  value: "товар #1: отсутствуют или пусты обязательные поля: quantity, sum"
                badJSON:
                  summary: Битый JSON
                  value: "некорректный JSON: unexpected end of JSON input"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /clear:
    get:
      tags: [window]
      summary: Очистить окно
      description: |
        Сбрасывает текст окна в `.` и переводит окно в полупрозрачный режим (alpha = 50).
      operationId: clearWindow
      responses:
        "200":
          description: Окно очищено
          content:
            text/plain:
              schema:
                type: string
              example: "Окно очищено и переведено в полупрозрачный режим"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /socials:
    get:
      tags: [window]
      summary: Показать QR-коды ботов на весь экран
      description: |
        Разворачивает окно на весь выбранный монитор, делает его полностью
        непрозрачным (alpha = 255) и показывает по центру QR-коды со ссылками
        из `telegram_url` / `max_url` (с подписями «Telegram» и «MAX»). Список
        покупок, текст и логотип скрываются.

        QR-коды берутся из `config.json` — если ссылки не заданы, экран останется
        пустым. Возврат к обычному режиму — через `POST /update` или `GET /clear`.
      operationId: showSocials
      responses:
        "200":
          description: Окно развёрнуто на весь экран, показаны QR-коды
          content:
            text/plain:
              schema:
                type: string
              example: "Окно развёрнуто на весь экран, показаны QR-коды"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /wifi:
    get:
      tags: [window]
      summary: Показать QR-код подключения к Wi-Fi на весь экран
      description: |
        Разворачивает окно на весь выбранный монитор, делает его полностью
        непрозрачным (alpha = 255) и показывает по центру вшитую картинку с
        QR-кодом подключения к Wi-Fi. Список покупок, текст и логотип скрываются.

        Возврат к обычному режиму — через `POST /update` или `GET /clear`.
      operationId: showWifi
      responses:
        "200":
          description: Окно развёрнуто на весь экран, показан QR-код Wi-Fi
          content:
            text/plain:
              schema:
                type: string
              example: "Окно развёрнуто на весь экран, показан QR-код Wi-Fi"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /shutdown:
    get:
      tags: [system]
      summary: Graceful shutdown
      description: |
        Останавливает HTTP-сервер (таймаут 5 секунд) и закрывает GUI-окно.
        Повторный вызов во время shutdown возвращает успех с другим сообщением.
      operationId: shutdown
      responses:
        "200":
          description: Shutdown инициирован или уже выполняется
          content:
            text/plain:
              schema:
                type: string
                enum:
                  - "Graceful shutdown запущен"
                  - "Graceful shutdown уже выполняется"
              examples:
                started:
                  summary: Первый вызов
                  value: "Graceful shutdown запущен"
                inProgress:
                  summary: Повторный вызов
                  value: "Graceful shutdown уже выполняется"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

  /docs:
    get:
      tags: [system]
      summary: Swagger UI
      description: |
        Редирект на встроенный Swagger UI (`/swagger/index.html`).
        Статические ресурсы UI отдаются из бинарника, спецификация — из `/openapi.yaml`.
      operationId: swaggerUI
      responses:
        "302":
          description: Редирект на Swagger UI
          headers:
            Location:
              schema:
                type: string
                example: /swagger/index.html

  /openapi.yaml:
    get:
      tags: [system]
      summary: OpenAPI-спецификация
      description: Возвращает встроенную OpenAPI 3.0 спецификацию этого API.
      operationId: getOpenAPISpec
      responses:
        "200":
          description: OpenAPI YAML
          content:
            application/yaml:
              schema:
                type: string
        "405":
          $ref: "#/components/responses/MethodNotAllowed"

components:
  schemas:
    PurchaseList:
      type: object
      required: [goods]
      properties:
        goods:
          type: array
          description: Список позиций чека (может быть пустым)
          items:
            $ref: "#/components/schemas/PurchaseItem"

    PurchaseItem:
      type: object
      required: [name, quantity, price, sum]
      properties:
        name:
          type: string
          description: Наименование товара
          example: "Молоко 3.2% 950мл"
        quantity:
          type: number
          description: Количество (целое или дробное)
          example: 2
        price:
          type: number
          description: Цена за единицу
          example: 89.90
        sum:
          type: number
          description: Сумма по позиции
          example: 179.80

    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
          enum: [ok]
          example: ok
        version:
          type: string
          description: Версия сборки (dev для локальных сборок, тег релиза для CI)
          example: v1.0.0

    MonitorInfo:
      type: object
      required: [index, left, top, right, bottom, width, height]
      properties:
        index:
          type: integer
          description: Порядковый индекс монитора (с 0)
          minimum: 0
          example: 0
        left:
          type: integer
          format: int32
          description: Левая координата прямоугольника монитора (px)
          example: 0
        top:
          type: integer
          format: int32
          description: Верхняя координата (px)
          example: 0
        right:
          type: integer
          format: int32
          description: Правая координата (px)
          example: 1920
        bottom:
          type: integer
          format: int32
          description: Нижняя координата (px)
          example: 1080
        width:
          type: integer
          format: int32
          description: Ширина монитора в пикселях
          example: 1920
        height:
          type: integer
          format: int32
          description: Высота монитора в пикселях
          example: 1080

  responses:
    MethodNotAllowed:
      description: HTTP-метод не поддерживается для данного эндпоинта
      content:
        text/plain:
          schema:
            type: string
          example: "Метод не поддерживается"

    InternalServerError:
      description: Внутренняя ошибка сервера
      content:
        text/plain:
          schema:
            type: string
          examples:
            readBody:
              summary: Ошибка чтения тела
              value: "Ошибка чтения тела запроса"
            monitors:
              summary: Ошибка списка мониторов
              value: "Не удалось получить список мониторов"